for file that misses can have a bad result

+++ entrypoint.sh
@@ -83,7 +83,7 @@ EOF
   } >>"$GITHUB_STEP_SUMMARY"
 
   # Surface "logged error ID" failures (if detected)
-  if [ -n "${MCIX_LOGGED_ERROR_ID:-}" ] [ -w "$GITHUB_STEP_SUMMARY" ]; then
+  if [ -n "${MCIX_LOGGED_ERROR_ID:-}" ] && [ -w "$GITHUB_STEP_SUMMARY" ]; then
     {
       echo "**❌ Error:** There was an error logged while running the command."
       if [ -n "${MCIX_LOGGED_ERROR_ID:-}" ]; then
@@ -142,7 +142,7 @@ EOF
 
   find / -name "cli.*.log" -ls
 
-  if [[ -f "${MCIX_LOG_DIR}/cli.$(date +%F).log" ]]; then
+  if [ -f "$MCIX_LOG_DIR/cli.$(date +%F).log" ]; then
     {
       # Display the contents of the mcix command's log file. (collapsed by default)
       echo '<details>'
@@ -157,18 +157,19 @@ EOF
       gh_warn "Log file for '${MCIX_CMD_NAME}' not found" "Continuing without failing the action."
   fi
 
-  for file in $MCIX_LOG_DIR/exception.*.log; do
-    echo " - $file"
-    {
-      # Display the contents of the mcix command's log file. (collapsed by default)
-      echo '<details>'
-      echo "<summary>Exception Log - $file</summary>"
-      echo # A blank line after the <summary> tag is required by GitHub to format the content correctly
-      echo '```'
-      cat $file
-      echo '```'
-      echo '</details>'
-    } >>"$GITHUB_STEP_SUMMARY"
+  for file in "$MCIX_LOG_DIR/exception.*.log"; do
+    if [ -f "$file" ]; then
+      {
+        # Display the contents of the mcix command's log file. (collapsed by default)
+        echo '<details>'
+        echo "<summary>Exception Log - $file</summary>"
+        echo # A blank line after the <summary> tag is required by GitHub to format the content correctly
+        echo '```'
+        cat $file
+        echo '```'
+        echo '</details>'
+      } >>"$GITHUB_STEP_SUMMARY"
+    fi
   done
 }
